fix: move links into Application layer to stick to dependency conventions.
